+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/replay/route/ReplayProgressObserver.kt
@@ -1,25 +1,30 @@
 package com.mapbox.navigation.core.replay.route
 
 import com.mapbox.api.directions.v5.models.RouteLeg
+import com.mapbox.geojson.Point
 import com.mapbox.navigation.base.trip.model.RouteProgress
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.replay.MapboxReplayer
+import com.mapbox.navigation.core.replay.history.ReplayEventBase
+import com.mapbox.navigation.core.replay.history.ReplayEventUpdateLocation
 import com.mapbox.navigation.core.trip.session.RouteProgressObserver
+import com.mapbox.turf.TurfConstants
+import com.mapbox.turf.TurfMeasurement
 
 /**
  * Register this to [MapboxNavigation.registerRouteProgressObserver].
  * This class will feed locations to your [MapboxReplayer] and simulate
  * your active route for you.
  */
-class ReplayProgressObserver(
+class ReplayProgressObserver @JvmOverloads constructor(
     /**
      * As navigation receives [RouteProgress], this will push events to your
      * replay history player to be played.
      */
-    private val mapboxReplayer: MapboxReplayer
+    private val mapboxReplayer: MapboxReplayer,
+    private val replayRouteMapper: ReplayRouteMapper = ReplayRouteMapper()
 ) : RouteProgressObserver {
 
-    private val replayRouteMapper = ReplayRouteMapper()
     private var currentRouteLeg: RouteLeg? = null
 
     /**
@@ -38,20 +43,53 @@ class ReplayProgressObserver(
      * @param routeProgress from the navigation session
      */
     override fun onRouteProgressChanged(routeProgress: RouteProgress) {
-        val routeProgressRouteLeg = routeProgress.currentLegProgress?.routeLeg
-        if (routeProgressRouteLeg != currentRouteLeg) {
+        val currentLegProgress = routeProgress.currentLegProgress
+        val routeProgressRouteLeg = currentLegProgress?.routeLeg
+        if (routeProgressRouteLeg != currentRouteLeg && currentLegProgress != null) {
             this.currentRouteLeg = routeProgressRouteLeg
-            onRouteLegChanged(routeProgressRouteLeg)
+            onRouteLegChanged(routeProgressRouteLeg, currentLegProgress.distanceTraveled)
         }
     }
 
-    private fun onRouteLegChanged(routeProgressRouteLeg: RouteLeg?) {
-        if (routeProgressRouteLeg != null) {
-            val replayEvents = replayRouteMapper.mapRouteLegGeometry(routeProgressRouteLeg)
-            if (replayEvents.isNotEmpty()) {
-                mapboxReplayer.pushEvents(replayEvents)
-                mapboxReplayer.seekTo(replayEvents.first())
+    private fun onRouteLegChanged(routeProgressRouteLeg: RouteLeg?, distanceTraveled: Float) {
+        if (routeProgressRouteLeg == null) return
+        val replayEvents = replayRouteMapper.mapRouteLegGeometry(routeProgressRouteLeg)
+        if (replayEvents.isEmpty()) return
+        val seekToIndex = indexAlong(replayEvents, distanceTraveled)
+        mapboxReplayer.pushEvents(replayEvents)
+        mapboxReplayer.seekTo(replayEvents[seekToIndex])
+    }
+
+    /**
+     * Measures the distance between the [events] and returns the index of an event that is greater
+     * than the [distanceTraveled]. If the distanceTraveled is greater than the event distance, the
+     * last index is returned. If the event list is empty -1 is returned.
+     *
+     * events: 0---1-------2------3--4-----5------
+     *         ==distanceTraveled====>||
+     * returns index: 5
+     */
+    private fun indexAlong(events: List<ReplayEventBase>, distanceTraveled: Float): Int {
+        var currentDistance = 0.0
+        var index = -1
+        var lastPoint: Point? = null
+        for (item in events) {
+            if (item is ReplayEventUpdateLocation) {
+                val currentPoint = Point.fromLngLat(item.location.lon, item.location.lat)
+                lastPoint?.let { point ->
+                    currentDistance += TurfMeasurement.distance(
+                        point,
+                        currentPoint,
+                        TurfConstants.UNIT_METERS
+                    )
+                }
+                lastPoint = currentPoint
+            }
+            index++
+            if (currentDistance >= distanceTraveled) {
+                return index
             }
         }
+        return index
     }
 }